requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pika==1.3.1
requests==2.28.2
django-cors-headers==4.0.0
pytest==7.3.1
pytest-django==4.5.2
gunicorn==20.1.0 